✨ feat: Add support for tool_choice in OpenAI API
Adds support for the `tool_choice` parameter in the OpenAI API. This allows users to specify a specific tool to use for a given completion.

The `tool_choice` parameter can be either a string or an object. If it is a string, it should be the name of the tool to use. If it is an object, it should have a `type` property set to `function` and a `function` property set to the name of the function to use.

This change also includes a number of other improvements to the OpenAI service, including:

- Improved error handling
- Better support for streaming completions
- More robust caching